feat(web): 👻 Ammo | Lucky Bag
- The Ammo component to display and perform some actions for the user bag was initialized.
- The Bag Button & Dialog were separated.
